+++ cmsplugin_blocks/cms_plugins.py
@@ -2,6 +2,8 @@
 """
 CMS Plugins installations
 """
+from django.conf import settings
+
 from cms.plugin_pool import plugin_pool
 
 from cmsplugin_blocks.plugins.album import AlbumPlugin
@@ -10,8 +12,15 @@ from cmsplugin_blocks.plugins.hero import HeroPlugin
 from cmsplugin_blocks.plugins.slider import SliderPlugin
 
 
-# TODO: May depend from a setting to only load some plugin
-plugin_pool.register_plugin(AlbumPlugin)
-plugin_pool.register_plugin(HeroPlugin)
-plugin_pool.register_plugin(CardPlugin)
-plugin_pool.register_plugin(SliderPlugin)
+# Register enabled plugins
+if "AlbumPlugin" in settings.BLOCKS_ENABLED_PLUGINS:
+    plugin_pool.register_plugin(AlbumPlugin)
+
+if "HeroPlugin" in settings.BLOCKS_ENABLED_PLUGINS:
+    plugin_pool.register_plugin(HeroPlugin)
+
+if "CardPlugin" in settings.BLOCKS_ENABLED_PLUGINS:
+    plugin_pool.register_plugin(CardPlugin)
+
+if "SliderPlugin" in settings.BLOCKS_ENABLED_PLUGINS:
+    plugin_pool.register_plugin(SliderPlugin)