+++ cmsplugin_blocks/models/card.py
@@ -19,28 +19,28 @@ class Card(SmartFormatMixin, CMSPlugin):
     A simple card object
     """
     ALIGNMENT_CHOICES = [
-        ('left', _('Content to the left, image to the right')),
-        ('right', _('Image to the left, content to the reft')),
+        ("left", _("Content to the left, image to the right")),
+        ("right", _("Image to the left, content to the reft")),
     ]
 
     alignment = models.CharField(
-        _('Alignment'),
+        _("Alignment"),
         choices=ALIGNMENT_CHOICES,
         max_length=15,
         blank=False,
-        default='left',
+        default="left",
     )
     template = models.CharField(
-        _('Template'),
+        _("Template"),
         blank=False,
         max_length=150,
         choices=get_card_template_choices(),
         default=get_card_default_template(),
-        help_text=_('Used template for content look.'),
+        help_text=_("Used template for content look."),
     )
     image = models.FileField(
-        _('Image'),
-        upload_to='blocks/card/%y/%m',
+        _("Image"),
+        upload_to="blocks/card/%y/%m",
         max_length=255,
         blank=True,
         null=True,
@@ -71,5 +71,5 @@ class Card(SmartFormatMixin, CMSPlugin):
         return self.media_format(self.image)
 
     class Meta:
-        verbose_name = _('Card')
-        verbose_name_plural = _('Cards')
+        verbose_name = _("Card")
+        verbose_name_plural = _("Cards")